+++ src/services/buscarUsuariosSACS.ts
@@ -1,25 +1,5 @@
 import { apiSACS } from '../api';
-
-interface UsuarioSACS {
-  codigo: number;
-  cpf: string;
-  rg: string;
-  nome: string;
-  cidade: string;
-  estado: string;
-  nmsexo: string;
-  naturalidade: string;
-  nacionalidade: string;
-  datanascimento: string;
-  telefone1: string;
-  telefone2: string;
-  email: string;
-  logradouro: string;
-  numero: number;
-  complemento: string;
-  bairro: string;
-  cep: string;
-}
+import { UsuarioSACS } from '../controllers/types';
 
 const buscarUsuariosSACS = async () => {
   const { data } = await apiSACS.get('/sacs-api');
